+++ src/evidently/ui/errors.py
@@ -1,9 +1,20 @@
+from starlette.responses import JSONResponse
+from starlette.responses import Response
+
+
 class EvidentlyServiceError(Exception):
-    pass
+    def to_response(self) -> Response:
+        raise NotImplementedError
 
 
 class EntityNotFound(EvidentlyServiceError):
     entity_name: str = ""
+
+    def to_response(self) -> Response:
+        return JSONResponse(
+            status_code=404,
+            content={"detail": f"{self.entity_name} not found"},
+        )
 
 
 class ProjectNotFound(EntityNotFound):
@@ -19,8 +30,16 @@ class UserNotFound(EntityNotFound):
 
 
 class NotEnoughPermissions(EvidentlyServiceError):
-    pass
+    def to_response(self) -> Response:
+        return JSONResponse(
+            status_code=403,
+            content={"detail": "Not enough permissions"},
+        )
 
 
 class NotAuthorized(EvidentlyServiceError):
-    pass
+    def to_response(self) -> Response:
+        return JSONResponse(
+            status_code=401,
+            content={"detail": "Not authorized"},
+        )
